requirements.txt
azure-ai-ml==1.11.0
azure-identity==1.15.0
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
mlflow==2.7.1
